+++ visualization/CHANGELOG.md
@@ -16,6 +16,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/)
 - Zooming doesn't affect the camera position [#3147](https://github.com/MaibornWolff/codecharta/pull/3665)
 - Fix the issue where the map disappears after excessive zooming out.[#3147](https://github.com/MaibornWolff/codecharta/pull/3665)
 
+### Added 🚀
+
+- Export multi-material 3MF for 3D printing of maps in color [#3616](https://github.com/MaibornWolff/codecharta/pull/3616)
+
 ### Changed
 
 - IMPORTANT: The sonar complexity will no longer be renamed to MCC! [#3456](https://github.com/MaibornWolff/codecharta/pull/3606)